fix: 20 bug fixes across security, correctness, build, docs — issues #1139-#1159
Mega-fix addressing all bugs found in adversarial audit:

Build: onebp_model library target, FLM conditional, always-defined macros
Security: OOB bounds in onebp_model, unchecked fread in gguf_reader/whisper/zamba2, ftell error in vulkan
Correctness: FP16-B dispatch, MoE tensor-shape detection, KV cache overflow, CPU_NUM_LAYERS empty
Docs: decode 2 base64-corrupted markdown files

+++ src/laguna_generate.cpp
@@ -178,8 +178,8 @@ struct DFlashSpec {
         printf("DFlash: %dL %dH %dV\n", L, H, V);
         
         k_cache.resize(L); v_cache.resize(L);
-        for(auto& k:k_cache) k.resize(max_pos*NKV*HD);
-        for(auto& v:v_cache) v.resize(max_pos*NKV*HD);
+        for(auto& k:k_cache) k.resize((size_t)max_pos * NKV * HD);
+        for(auto& v:v_cache) v.resize((size_t)max_pos * NKV * HD);
         return true;
     }
     
@@ -305,8 +305,8 @@ struct LagunaInference {
         printf("Laguna: %dL %dH %dV %dE/%d\n", L, H, V, NE, NEU);
         
         k_cache.resize(L); v_cache.resize(L);
-        for(auto& k:k_cache) k.resize(max_pos*NKV*HD);
-        for(auto& v:v_cache) v.resize(max_pos*NKV*HD);
+        for(auto& k:k_cache) k.resize((size_t)max_pos * NKV * HD);
+        for(auto& v:v_cache) v.resize((size_t)max_pos * NKV * HD);
         return true;
     }
